+++ tests/hazmat/primitives/test_ec.py
@@ -68,10 +68,9 @@ class DummyCurve(object):
 
 @utils.register_interface(interfaces.EllipticCurveSignatureAlgorithm)
 class DummySignatureAlgorithm(object):
-    pass
+    algorithm = None
 
 
-@utils.register_interface(EllipticCurveBackend)
 class DeprecatedDummyECBackend(object):
     def elliptic_curve_private_key_from_numbers(self, numbers):
         return b"private_key"
@@ -80,7 +79,7 @@ class DeprecatedDummyECBackend(object):
         return b"public_key"
 
 
-@pytest.mark.elliptic
+@pytest.mark.requires_backend_interface(interface=EllipticCurveBackend)
 def test_skip_curve_unsupported(backend):
     with pytest.raises(pytest.skip.Exception):
         _skip_curve_unsupported(backend, DummyCurve())
@@ -138,7 +137,7 @@ def test_ec_numbers():
         )
 
 
-@pytest.mark.elliptic
+@pytest.mark.requires_backend_interface(interface=EllipticCurveBackend)
 class TestECWithNumbers(object):
     @pytest.mark.parametrize(
         ("vector", "hash_type"),
@@ -174,7 +173,7 @@ class TestECWithNumbers(object):
             assert curve_type().name == priv_num.public_numbers.curve.name
 
 
-@pytest.mark.elliptic
+@pytest.mark.requires_backend_interface(interface=EllipticCurveBackend)
 class TestECDSAVectors(object):
     @pytest.mark.parametrize(
         ("vector", "hash_type"),
@@ -268,6 +267,28 @@ class TestECDSAVectors(object):
             ec.EllipticCurvePublicNumbers(
                 47250808410327023131573602008345894927686381772325561185532964,
                 1120253292479243545483756778742719537373113335231773536789915,
+                ec.SECP256R1(),
+            )
+        )
+        with pytest.raises(ValueError):
+            numbers.private_key(backend)
+
+        numbers = ec.EllipticCurvePrivateNumbers(
+            357646505660320080863666618182642070958081774038609089496899025506,
+            ec.EllipticCurvePublicNumbers(
+                -4725080841032702313157360200834589492768638177232556118553296,
+                1120253292479243545483756778742719537373113335231773536789915,
+                ec.SECP256R1(),
+            )
+        )
+        with pytest.raises(ValueError):
+            numbers.private_key(backend)
+
+        numbers = ec.EllipticCurvePrivateNumbers(
+            357646505660320080863666618182642070958081774038609089496899025506,
+            ec.EllipticCurvePublicNumbers(
+                47250808410327023131573602008345894927686381772325561185532964,
+                -1120253292479243545483756778742719537373113335231773536789915,
                 ec.SECP256R1(),
             )
         )