+++ tests/utils.py
@@ -40,8 +40,6 @@ def select_backends(names, backend_list):
     if names is None:
         return backend_list
     split_names = [x.strip() for x in names.split(',')]
-    # this must be duplicated and then removed to preserve the metadata
-    # pytest associates. Appending backends to a new list doesn't seem to work
     selected_backends = []
     for backend in backend_list:
         if backend.name in split_names:
@@ -55,12 +53,13 @@ def select_backends(names, backend_list):
         )
 
 
-def check_for_iface(name, iface, item):
-    if name in item.keywords and "backend" in item.funcargs:
-        if not isinstance(item.funcargs["backend"], iface):
-            pytest.skip("{0} backend does not support {1}".format(
-                item.funcargs["backend"], name
-            ))
+def skip_if_empty(backend_list, required_interfaces):
+    if not backend_list:
+        pytest.skip(
+            "No backends provided supply the interface: {0}".format(
+                ", ".join(iface.__name__ for iface in required_interfaces)
+            )
+        )
 
 
 def check_backend_support(item):